+++ apps/multichoice/lib/layouts/home_layout/card_layout.dart
@@ -22,7 +22,7 @@ class EntryLayout extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return context.watch<AppLayout>().appLayout
+    return context.watch<AppLayout>().isLayoutVertical
         ? _VerticalEntry(
             id: id,
             entries: entries,