+++ packages/models/lib/src/enums/storage/storage_keys.dart
@@ -1,6 +1,8 @@
 enum StorageKeys {
+  isDarkMode('_isDarkMode'),
   currentStep('_productTourCurrentStep'),
-  isCompleted('_productTourIsCompleted');
+  isCompleted('_productTourIsCompleted'),
+  isLayoutVertical('_isLayoutVertical');
 
   const StorageKeys(this.key);